\name{clustal}
     \docType{data}
     \alias{clustal}
     \title{Example of results obtained after a call to read.alignment}
     \description{This data set gives an example of a protein alignment obtained after a call to the function read.alignment on an alignment file in "clustal" format.}
     \usage{data(clustal)}
     \format{A List of class alignment}
     \source{http://www.clustal.org/}
     \references{
      Thompson, J.D., Higgins D.G., Gibson T.J. (1994) \emph{CLUSTAL W: improving the sensitivity of progressive multiple sequence alignment through sequence weighting, position specific gap penalties and weight matrix choice}.
      Nucleic Acids Res. 22(22):4673-80.
     }
     \keyword{datasets}
