\name{fasta}
     \docType{data}
     \alias{fasta}
     \title{Example of results obtained after a call to read.alignment}
     \description{This data set gives an example of a amino acids alignment obtained after a call to the function read.alignment on an alignment file in "fasta" format.}
     \usage{data(fasta)}
     \format{A List of class alignment}
     \source{https://pbil.univ-lyon1.fr/help/formats.html/}
     \references{Pearson W.R. and Lipman D.J. (1988) \emph{Improved tools for biological sequence comparison.}.Proc Natl Acad Sci U S A. 85(8):2444-8.}
     \keyword{datasets}
