\name{invers}
\alias{invers}
\title{invers}
\description{
 reverse a sequence.
}
\usage{
invers(seq)
}
\arguments{
  \item{seq}{\code{seq} is a vector of character}
}
\value{
 return a vector of character.
}
\references{\url{charif@biomserv.univ-lyon1.fr}}
\author{delphine charif}
\examples{
s=c("A","G","T","A","C","G")
invers(s)
}
\keyword{manip}

