\name{lseqinr}
\alias{lseqinr}
\title{ To see what's inside the package seqinr }
\description{
  This is just a shortcut for ls("package:seqinr")
}
\usage{
lseqinr()
}
\value{
  The list of objects in the package seqinr  
}
\references{ 
To have an overview of the seqinR's functionnality, please consult this vignette: 
Charif, D., Lobry, J.R. (2005) SeqinR: a contributed package to the R project for statistical
computing devoted to biological sequences retrieval and analysis. Springer Verlag, \emph{Biological and Medical Physics/Biomedical Series}, in preparation. \cr
}
\note{ 
Use \code{library(help=seqinr)} to have a summary of the functionc available in the package.
}
\author{
J.R. Lobry 
}
\examples{
lseqinr()
}
\keyword{utilities}
