\name{mase}
     \docType{data}
     \alias{mase}
     \title{Example of results obtained after a call to read.alignment}
     \description{This data set gives an example of a protein alignment obtained after a call to the function read.alignment on an alignment file in "mase" format.}
     \usage{data(mase)}
     \format{A List of class alignment}
     \source{http://www.clustal.org/}
     \references{Faullcner.D.V. and Jurka,J. (1988) \emph{Multiple sequences alignment editor(MASE).} Trends Biochem. Sa., 13, 321-322.}
     \keyword{datasets}
