\name{msf}
     \docType{data}
     \alias{msf}
     \title{Example of results obtained after a call to read.alignment}
     \description{This data set gives an example of a protein alignment obtained after a call to the function read.alignment on an alignment file in "msf" format.}
     \usage{data(msf)}
     \format{A List of class alignment}
     \source{http://www.ebi.ac.uk/2can/tutorials/formats.html#MSF/}
     \keyword{datasets}
