\name{phylip}
     \docType{data}
     \alias{phylip}
     \title{Example of results obtained after a call to read.alignment}
     \description{This data set gives an example of a amino acids alignment obtained after a call to the function read.alignment on an alignment file in "phylip" format.}
     \usage{data(phylip)}
     \format{A List of class alignment}
     \source{http://evolution.genetics.washington.edu/phylip.html}
     \references{http://evolution.genetics.washington.edu/phylip.html}
     \keyword{datasets}
