\name{revaligntest}
\alias{revaligntest}
\docType{data}
\title{ Three aligned nucleic acid sequences }
\encoding{latin1}
\description{
This dataset is used as a sanity check in \code{\link{reverse.align}}.
}
\usage{
data(revaligntest)
}
\format{
  An object of class alignment with 3 sequences.
}
\references{
  \code{citation("seqinr")}
}
\seealso{ \code{\link{reverse.align}} }
\examples{
data(revaligntest)
}
\keyword{ datasets }
