\name{syncodons}
\alias{syncodons}
\title{Synonymous codons}
\description{
  Returns all synonymous codons for each codon given
}
\usage{syncodons(codons, numcode = 1)}
\arguments{
  \item{codons}{ A sequence of codons as generated by \code{splitseq} }
  \item{numcode}{ The genetic code number as in \code{translate} }
}
\value{
  a list containing, for each codon given (list tags), all synonymous
  codons (including the original one)
}
\references{
  \code{citation("seqinr")}
}
\author{ Leonor Palmeira }
\seealso{ \code{\link{synsequence}} }
\examples{
  syncodons(words(3))
}
\keyword{ utilities }
