\name{toyaa}
\alias{toyaa}
\docType{data}
\title{A toy example of amino-acid counts in three proteins}
\encoding{latin1}
\description{
This is a toy data set to illustrate the importance of metric choice.
}
\usage{data(toyaa)}
\format{
  A data frame with 3 observations on the following 3 variables:
  \describe{
    \item{Ala}{Alanine counts}
    \item{Val}{Valine counts}
    \item{Cys}{Cysteine counts}
  }
}
\source{
This toy example was inspired by Gautier, C:
Analyses statistiques et évolution des séquences d'acides nucléiques.
PhD thesis (1987), Université Claude Bernard - Lyon I.
}
\references{
  \code{citation("seqinr")}
}
\examples{
data(toyaa)
}
\keyword{datasets}
