\name{toycodon}
\alias{toycodon}
\docType{data}
\title{A toy example of codon counts in three coding sequences}
\description{
This is a toy data set to illustrate synonymous and non-synonymous codon usage analyses.
}
\usage{data(toyaa)}
\format{
  A data frame with 3 observations (coding sequences) for 10 codons.}
\source{
Created for release 1.0-4 of seqinr's vignette.
}
\references{
  \code{citation("seqinr")}
}
\examples{
data(toycodon)
}
\keyword{datasets}
