\name{lseqinr}
\alias{lseqinr}
\title{ To see what's inside the package seqinr }
\description{
  This is just a shortcut for \code{ls("package:seqinr")}
}
\usage{
lseqinr()
}
\value{
  The list of objects in the package \code{seqinr}  
}
\references{
  \code{citation("seqinr")}
}
\note{ 
Use \code{library(help=seqinr)} to have a summary of the functions available in the package.
}
\author{J.R. Lobry}
\examples{
lseqinr()
}
\keyword{utilities}
