\name{seqinr-package}
\alias{seqinr-package}
\alias{seqinr}
\docType{package}
\title{
Biological Sequences Retrieval and Analysis
}
\description{
Exploratory data analysis and data visualization for
biological sequence (DNA and protein) data. Include also
utilities for sequence data management under the ACNUC system.
}
\author{
Delphine Charif and Jean R. Lobry and Anamaria Necsulea and Leonor Palmeira

Maintainer: Simon Penel <penel@biomserv.univ-lyon1.fr>
}
\references{
citation('seqinr')
}
\keyword{ package }

