\name{seqinr-package}
\alias{seqinr-package}
\alias{seqinr}
\docType{package}
\title{
Biological Sequences Retrieval and Analysis
}
\description{
Exploratory data analysis and data visualization for
biological sequence (DNA and protein) data. Include also
utilities for sequence data management under the ACNUC system.
}
\author{Delphine Charif [aut], Olivier Clerc [ctb], Carolin Frank [ctb], 
  Jean R. Lobry [aut], Anamaria Necşulea [ctb], Leonor Palmeira [ctb], 
  Simon Penel [cre], Guy Perrière [ctb]}
\references{
citation('seqinr')
}
\keyword{ package }

